Updated RTC, SysTick, and delay to ensure that all over flow logic and conditions are defined for the various time keeping methods

+++ src/delay.h
@@ -26,9 +26,9 @@
 extern "C" {
 #endif
 
-uint64_t millis();
+uint32_t millis();
 void     delay( uint32_t ms );
-uint64_t micros();
+uint32_t micros();
 void     delayMicroseconds( uint32_t us );
 
 #ifdef __cplusplus
